Correct plugin order
The order in which key events are processed by plugins is very
important. In particular, there are several plugins that change the
value of a key, or delay or abort key events entirely, and these
plugins really need to come before the plugins that act on those key
values. Specifically, Qukeys really needs to come before most other
plugins that have `onKeyswitchEvent()` handlers.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>