Enable additional plugins for sketches where we can
For sketches where we can, enable additional - non-intrusive - plugins. In every
case care has been taken to not loose any functionality with the new plugins,
they're strictly additions.

This means that some sketches are less featureful than their former
"experimental" variants were.

In particular, the Keyboardio Atreus and Keyboardio Model01 firmwares are
affected. The latter is left as-is, without any changes, because it's already
tight. The former had DynamicMacros taken out compared to the previous
experimental firmware, because of EEPROM space concerns.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>