Makefile: Do the version mangling in the Makefile
Instead of setting EXTRA_VERSION to `.N` or `+N` depending on which workflow our
CI is running, figure all that out in the Makefile. The `version` target will
print the proper version string for both snapshot and tagged releases. The new
`version-tag` target will print the tag name to be used in CI.

This lets us simplify and merge our workflows.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>