Use the firmware from Kaleidoscope as upstream sources
The initial idea behind this repo was that this will serve as the canonical
source of firmware. That did not came to be, because we ended up syncing a copy
back to Kaleidoscope.

To remedy the situation, we came to the decision that we'll use Kaleidoscope as
the upstream for the firmware. The firmware here were already synced back there,
so we can now pull Kaleidoscope in as a submodule, and just symlink the
sketches.

We will still keep using this repository, to publish the artifacts, and to have
better control over which version of Kaleidoscope we build against.

Fixes #37.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>